ContingencyTable: Moved row/column/all selection logic from OWConfusionMatrix
